backend/requirements.txt
absl-py==1.4.0
acme==1.21.0
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.15
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.9.1.post1
apturl==0.5.2
asgiref==3.5.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.2.0
Babel==2.14.0
bcrypt==3.2.0
bidict==0.23.1
blinker==1.4
boto3==1.34.88
botocore==1.34.88
Brlapi==0.8.3
CacheControl==0.12.14
cached-property==1.5.2
cachetools==5.3.3
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
click==8.0.3
cloudpickle==2.2.1
colorama==0.4.4
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
comm==0.2.2
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
confluent-kafka==2.3.0
cryptography==42.0.5
cupshelpers==1.0
cycler==0.12.1
dask==2022.10.2
dbus-python==1.2.18
debugpy==1.8.5
decorator==5.1.1
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.12
dnspython==2.3.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
exceptiongroup==1.2.0
executing==2.1.0
fasteners==0.14.1
fbmessenger==6.0.0
filelock==3.13.3
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
future==0.18.2
gast==0.4.0
git-filter-repo==2.34.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.2
gyp==0.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.27.0
jax==0.4.26
jedi==0.19.1
jeepney==0.7.1
jmespath==1.0.1
joblib==1.2.0
josepy==1.10.0
jsonpickle==3.0.4
jsonschema==3.2.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kazam==1.4.5
keras==2.12.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
locket==1.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mattermostwrapper==2.2
ml-dtypes==0.4.0
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.8
multidict==5.2.0
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==2.6.3
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
openai==1.19.0
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
pandas==2.2.2
paramiko==2.9.3
parsedatetime==2.6
parso==0.8.4
partd==1.4.1
pexpect==4.8.0

platformdirs==4.2.0
pluggy==1.5.0
portalocker==2.8.2
prompt_toolkit==3.0.47
protobuf==4.23.3
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==1.10.9
pydantic_core==2.18.1
pydot==1.4.2
Pygments==2.18.0
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pykwalify==1.8.0
pymacaroons==0.13.0
pymongo==4.3.3
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==26.2.0
questionary==1.10.0
randomname==0.1.5
rasa==3.6.20
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rocketchat-API==1.30.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentry-sdk==1.14.0
simple-websocket==1.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack_sdk==3.27.1
sniffio==1.3.1
SQLAlchemy==1.4.52
sqlparse==0.4.2
ssh-import-id==5.11
stack-data==0.6.3
structlog==23.3.0
structlog-sentry==2.1.0
supervisor==4.2.1
systemd-python==234
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.12.0
termcolor==2.4.0
terminaltables==3.1.10
texttable==1.6.4
threadpoolctl==3.4.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
twilio==8.2.2
typing-utils==0.1.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.9.0
unattended-upgrades==0.1
urllib3==2.2.1
usb-creator==0.3.7
uvloop==0.19.0
virtualenv==20.25.1
wadllib==1.3.6
wcwidth==0.2.13
webexteamssdk==1.6.1
websocket-client==1.2.3
websockets==10.4
Werkzeug==3.0.2
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0


djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
django-filter==23.2
psycopg2-binary==2.9.6
Pillow==9.5.0
python-dotenv==1.0.0